requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
adjusttext=0.8=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.11.1=pypi_0
anndata=0.8.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
arrow=1.2.3=pypi_0
asciitree=0.3.3=pypi_0
astroid=3.0.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
backoff=2.2.1=pypi_0
bbknn=1.6.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
binaryornot=0.4.4=pypi_0
bleach=6.0.0=pypi_0
blessed=1.20.0=pypi_0
bokeh=3.2.2=pypi_0
boltons=23.0.0=pypi_0
boto3=1.34.3=pypi_0
botocore=1.34.3=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2023.7.22=pypi_0
chardet=5.1.0=pypi_0
charset-normalizer=2.1.1=pypi_0
chex=0.1.7=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
cmaes=0.10.0=pypi_0
colorama=0.4.6=pypi_0
colorcet=3.0.1=pypi_0
colorlog=6.7.0=pypi_0
comm=0.2.0=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.1.1=pypi_0
cookiecutter=2.1.1=pypi_0
cramjam=2.7.0=pypi_0
croniter=1.4.1=pypi_0
crossref-commons=0.0.7=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
dask=2022.2.0=pypi_0
datacache=1.2.1=pypi_0
datashader=0.14.4=pypi_0
datashape=0.5.2=pypi_0
dateutils=0.6.12=pypi_0
dcor=0.5.3=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
deepdiff=6.7.1=pypi_0
descriptastorus=2.6.0=pypi_0
dgl=1.0.1=pypi_0
dgllife=0.3.2=pypi_0
dill=0.3.7=pypi_0
dm-tree=0.1.8=pypi_0
dnspython=2.3.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docrep=0.3.2=pypi_0
docutils=0.19=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
etils=1.6.0=pypi_0
exceptiongroup=1.1.1=pypi_0
fastapi=0.105.0=pypi_0
fasteners=0.18=pypi_0
fastparquet=2023.8.0=pypi_0
filelock=3.12.4=pypi_0
flatbuffers=23.5.26=pypi_0
flatten-dict=0.4.2=pypi_0
flax=0.7.5=pypi_0
fonttools=4.38.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.1.0=pypi_0
future=0.18.3=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=2.0.2=pypi_0
grpcio=1.53.0=pypi_0
gtfparse=1.3.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.8.0=pypi_0
holoviews=1.16.0=pypi_0
hyperopt=0.1.2=pypi_0
idna=3.4=pypi_0
imageio=2.28.1=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
inquirer=3.1.4=pypi_0
ipython=7.34.0=pypi_0
ipywidgets=8.1.1=pypi_0
isort=5.12.0=pypi_0
itsdangerous=2.1.2=pypi_0
jax=0.4.23=pypi_0
jaxlib=0.4.23=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
jinja2-time=0.2.0=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jsonpickle=3.0.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keras=2.15.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
levenshtein=0.23.0=pypi_0
libclang=16.0.0=pypi_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_1
libgomp=13.2.0=h807b86a_1
libnsl=2.0.0=h7f98852_0
libsqlite=3.43.0=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=hd590300_5
lightning=2.0.9.post0=pypi_0
lightning-cloud=0.5.57=pypi_0
lightning-utilities=0.9.0=pypi_0
linkify-it-py=2.0.2=pypi_0
llvmlite=0.40.1=pypi_0
locket=1.0.0=pypi_0
loompy=3.0.7=pypi_0
mako=1.2.4=pypi_0
markdown=3.4.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.7.0=pypi_0
mdit-py-plugins=0.4.0=pypi_0
mdurl=0.1.2=pypi_0
memoized-property=1.0.3=pypi_0
ml-collections=0.1.1=pypi_0
ml-dtypes=0.3.1=pypi_0
mock=5.1.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.7=pypi_0
mudata=0.2.3=pypi_0
multidict=6.0.4=pypi_0
multipledispatch=1.0.0=pypi_0
munch=4.0.0=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.8=pypi_0
networkx=3.1=pypi_0
nose=1.3.7=pypi_0
numba=0.57.1=pypi_0
numcodecs=0.12.1=pypi_0
numpy=1.24.4=pypi_0
numpy-groupies=0.10.2=pypi_0
numpyro=0.13.2=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.2.140=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
obonet=1.0.0=pypi_0
omegaconf=2.3.0=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.1.2=hd590300_0
opt-einsum=3.3.0=pypi_0
optax=0.1.7=pypi_0
optuna=3.3.0=pypi_0
orbax-checkpoint=0.4.8=pypi_0
ordered-set=4.1.0=pypi_0
owlready2=0.45=pypi_0
packaging=23.1=pypi_0
pandas=2.1.1=pypi_0
pandas-flavor=0.6.0=pypi_0
panel=1.2.3=pypi_0
param=1.13.0=pypi_0
parso=0.8.3=pypi_0
partd=1.4.0=pypi_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=pyhd8ed1ab_0
platformdirs=4.0.0=pypi_0
plotly=5.18.0=pypi_0
pluggy=1.3.0=pypi_0
progressbar33=2.4=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
py4j=0.10.9.7=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyct=0.5.0=pypi_0
pydantic=2.1.1=pypi_0
pydantic-core=2.4.0=pypi_0
pyensembl=2.2.9=pypi_0
pygments=2.16.1=pypi_0
pyjwt=2.8.0=pypi_0
pylint=3.0.2=pypi_0
pymongo=4.6.1=pypi_0
pynndescent=0.5.10=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.1=pypi_0
pyro-api=0.1.2=pypi_0
pyro-ppl=1.8.6=pypi_0
pytest=7.4.3=pypi_0
python=3.11.4=hab00c5b_0_cpython
python-annoy=1.17.2=py311hb755f60_1
python-dateutil=2.8.2=pypi_0
python-editor=1.0.4=pypi_0
python-levenshtein=0.23.0=pypi_0
python-multipart=0.0.6=pypi_0
python-slugify=8.0.1=pypi_0
python_abi=3.11=4_cp311
pytorch-lightning=2.1.2=pypi_0
pytz=2023.3.post1=pypi_0
pyviz-comms=3.0.0=pypi_0
pyyaml=6.0.1=pypi_0
questionary=2.0.1=pypi_0
rapidfuzz=3.5.2=pypi_0
ratelimit=2.2.1=pypi_0
rdkit=2023.3.3=pypi_0
readchar=4.0.5=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.0=pypi_0
rsa=4.9=pypi_0
s3transfer=0.9.0=pypi_0
sacred=0.8.5=pypi_0
scanpy=1.9.5=pypi_0
scikit-learn=1.3.0=pypi_0
scikit-misc=0.3.1=pypi_0
scipy=1.11.2=pypi_0
scvi=0.6.8=pypi_0
scvi-tools=1.0.4=pypi_0
seaborn=0.13.0=pypi_0
seml=0.4.0=pypi_0
serializable=0.2.1=pypi_0
session-info=1.0.0=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
sfaira=0.3.12=pypi_0
simplejson=3.19.2=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sparse=0.14.0=pypi_0
sqlalchemy=2.0.20=pypi_0
starlette=0.27.0=pypi_0
starsessions=1.3.0=pypi_0
statsmodels=0.14.0=pypi_0
stdlib-list=0.9.0=pypi_0
switchlang=0.1.0=pypi_0
sympy=1.12=pypi_0
tbb=2021.10.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0.post1=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
tensorstore=0.1.51=pypi_0
termcolor=2.4.0=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tinytimer=0.0.0=pypi_0
tk=8.6.12=h27826a3_0
tomlkit=0.12.3=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.0=pypi_0
torch-geometric=2.3.1=pypi_0
torchmetrics=1.2.0=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.10.0=pypi_0
triton=2.1.0=pypi_0
typechecks=0.1.0=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
umap-learn=0.5.4=pypi_0
urllib3=2.0.5=pypi_0
uvicorn=0.24.0.post1=pypi_0
versioneer=0.29=pypi_0
vscodedebugvisualizer=0.1.0=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.14.1=pypi_0
xarray=2023.8.0=pypi_0
xlrd=1.2.0=pypi_0
xyzservices=2023.7.0=pypi_0
xz=5.2.6=h166bdaf_0
yarl=1.9.4=pypi_0
zarr=2.16.1=pypi_0
zipp=3.17.0=pypi_0
